Finish refactoring enemies
