+++ Assets/Prefabs/Enemy/EnemyPhantom.prefab
@@ -11,7 +11,7 @@ GameObject:
   - component: {fileID: 8126026953933692042}
   - component: {fileID: 872007728857571396}
   - component: {fileID: 2089366822705537239}
-  - component: {fileID: 7741952854358576496}
+  - component: {fileID: -3960653538617794907}
   m_Layer: 11
   m_Name: EnemyPhantom
   m_TagString: Enemy
@@ -97,7 +97,7 @@ CircleCollider2D:
   m_Offset: {x: -0.0025580227, y: 0}
   serializedVersion: 2
   m_Radius: 0.13804212
---- !u!114 &7741952854358576496
+--- !u!114 &-3960653538617794907
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -106,13 +106,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 712027675066443393}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4d91f7b52cec78c40b33c6df5ebb84ca, type: 3}
+  m_Script: {fileID: 11500000, guid: 93af27ec3f8342a469e6d88c0a16b361, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  speed: 3
+  speed: 2
   health: 3
   attackDamage: 1
-  attackRange: 0.29
+  attackRange: 0.5
   attackCooldown: 1
 --- !u!1 &6640638025385805139
 GameObject:
