+++ Assets/Scenes/SampleScene.unity
@@ -154,8 +154,6 @@ MonoBehaviour:
   uiPrefab: {fileID: 0}
   killLimit: 5
   totalKills: 0
-  maxTowerCount: 0
-  curTowerCount: 0
   maxWorkerCount: 0
   curWorkerCount: 0
   shouldSpawnMine: 1
@@ -513,7 +511,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &246233309
 Transform:
   m_ObjectHideFlags: 0
@@ -4107,7 +4105,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!114 &1844260905
 MonoBehaviour:
   m_ObjectHideFlags: 0
