+++ Assets/Scripts/EnemyScripts/EnemyAbstract.cs
@@ -1,9 +1,10 @@
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using System;
+using UnityEditor.Experimental.GraphView;
 
-public class EnemyAbstract : MonoBehaviour, IDamageable
+public abstract class EnemyAbstract : MonoBehaviour, IDamageable
 {
     public static event Action enemyKill;
 
@@ -17,7 +18,7 @@ public class EnemyAbstract : MonoBehaviour, IDamageable
     public float attackCooldown = 1f; // Attack cooldown time
     protected bool canAttack = true;
     protected Rigidbody2D rb;
-    protected string name = "Please rename me in StartCall";
+    protected string enemyType = "Please rename me in StartCall";
     
     // do not overwrite
     private void OnDisable()
@@ -28,8 +29,16 @@ public class EnemyAbstract : MonoBehaviour, IDamageable
     // do not overwrite
     void Start()
     {
-        core = GameObject.FindGameObjectWithTag("Core").transform;
-        target = core; // Initial target is the Core
+        GameObject coreObject = GameObject.FindGameObjectWithTag("Core");
+        if (coreObject != null)
+        {
+            core = coreObject.transform;
+            target = core; // Initial target is the Core
+        }
+        else
+        {
+            Debug.LogError("BossEnemy: 找不到 `Core`，请检查是否存在带 `Core` 标签的对象！"); // no idea what in the debug XD
+        }
         Physics2D.IgnoreLayerCollision(LayerMask.NameToLayer("EnemyEnCol"), LayerMask.NameToLayer("EnemyEnCol"));
         Physics2D.IgnoreLayerCollision(LayerMask.NameToLayer("EnemyEnCol"), LayerMask.NameToLayer("Hero"));
 
@@ -39,19 +48,50 @@ public class EnemyAbstract : MonoBehaviour, IDamageable
         StartCall();
     }
 
-    // override for addition starting setting
-    protected abstract void StartCall();
-    // **Set aggro target**
-    public abstract void SetAggroTarget(Transform newTarget);
+    // **Take damage**
+    public void TakeDamage(int damage, Transform attacker)
+    {
+        health -= damage;
+        if (health <= 0)
+        {
+            Debug.Log(enemyType + " enemy dies");
+            Destroy(gameObject); // Destroy the enemy when health reaches zero
+        }
+        else
+        {
+            PostDamage(damage, attacker);
+        }
+    }
 
     // do not overwrite
     void Update()
     {
+        BeforeUpdate();
         if (target == null) return;
 
         Move();
         TryAttack();
     }
+
+    // **Return to Core if the target is destroyed**
+    protected virtual void LateUpdate()
+    {
+        if (isAggroed && target == null)
+        {
+            target = core;
+            isAggroed = false;
+        }
+    }
+
+    protected virtual void PostDamage(int damage, Transform attacker)
+    {
+        SetAggroTarget(attacker); // Switch target to the attacker after taking damage
+    }
+
+    // override for addition starting setting
+    protected abstract void StartCall();
+    // **Set aggro target**
+    public abstract void SetAggroTarget(Transform newTarget);
 
     // how the enemey moves. default is to move toward target
     protected virtual void Move()
@@ -68,8 +108,14 @@ public class EnemyAbstract : MonoBehaviour, IDamageable
         }
     }
 
+    // Any setting before executing update
+    protected virtual void BeforeUpdate()
+    {
+
+    }
+
     // **Attack logic**
-    IEnumerator AttackTarget()
+    protected virtual IEnumerator AttackTarget()
     {
         canAttack = false;
         if (target.GetComponent<Health>() != null)
@@ -78,30 +124,5 @@ public class EnemyAbstract : MonoBehaviour, IDamageable
         }
         yield return new WaitForSeconds(attackCooldown);
         canAttack = true;
-    }
-
-    // **Take damage**
-    public void TakeDamage(int damage, Transform attacker)
-    {
-        health -= damage;
-        if (health <= 0)
-        {
-            Debug.Log("enemy dies");
-            Destroy(gameObject); // Destroy the enemy when health reaches zero
-        }
-        else
-        {
-            SetAggroTarget(attacker); // Switch target to the attacker after taking damage
-        }
-    }
-
-    // **Return to Core if the target is destroyed**
-    void LateUpdate()
-    {
-        if (isAggroed && target == null)
-        {
-            target = core;
-            isAggroed = false;
-        }
     }
 }
